# api.yaml
openapi: 3.0.0
info:
  title: Book API
  version: 1.0.0
  description: A simple API for managing books.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /books:
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - year
              properties:
                title:
                  type: string
                  example: "To Kill a Mockingbird"
                author:
                  type: string
                  example: "Harper Lee"
                year:
                  type: integer
                  example: 1960
      responses:
        201:
          description: Book created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  title:
                    type: string
                  author:
                    type: string
                  year:
                    type: integer
        400:
          description: Bad Request - Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      summary: Get all books
      responses:
        200:
          description: A list of all books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    title:
                      type: string
                    author:
                      type: string
                    year:
                      type: integer

  /books/{id}:
    get:
      summary: Get a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: A single book
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  author:
                    type: string
                  year:
                    type: integer
        404:
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    put:
      summary: Update a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - year
              properties:
                title:
                  type: string
                  example: "Updated Title"
                author:
                  type: string
                  example: "Updated Author"
                year:
                  type: integer
                  example: 2021
      responses:
        200:
          description: Book updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  author:
                    type: string
                  year:
                    type: integer
        400:
          description: Bad Request - Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        204:
          description: Book deleted successfully
        404:
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string